refactor: refactoring assets structure folder
