+++ src/frontend/src/index.css
@@ -9,7 +9,7 @@
 
 @font-face {
   font-family: "Lexend";
-  src: url("./assets/Lexend.ttf") format("truetype");
+  src: url("./assets/fonts/Lexend.ttf") format("truetype");
   font-weight: 400;
   font-style: normal;
 }
