+++ src/frontend/src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import RootLayout from '@/components/RootLayout';
-import logo from '@/assets/eduChain.svg';
+import logo from '@/assets/icons/eduChain.svg';
 import { Link } from 'react-router-dom';
 
 export default function Home() {
